refactor: extract long-press reorder gesture into useLongPressReorder (#745)
The ~200-line pointer state machine (hold threshold, move tolerance,
hit-testing, iOS synthetic-click suppression, non-passive scroll
blocking) moves to a reusable composable. WorkoutTracker supplies the
policy via options: filtered-list disable, the + Log ignore zone, the
haptic, and the filtered-index -> store-index mapping in onReorder.
Keyboard reorder stays in the component (store-coupled, shares no
gesture state). Behavior is pinned by the existing long-press gesture
test suite.
